colorizer/requirements.txt


Pillow>=10.0.0
pystray>=0.19.5
pywin32>=306


